Initial Commit
- Added Vagrantfile
- Added build.sh
- Added sun5i-r8-chip.dtb

I will add a formal readme later. The gist is that Vagrantfile is a configuration for Vagrant-Libvirt that emulates the ARM Hardware in the provided device tree blob using QEMU.
The build.sh is a script for bootstrapping and compiling a kernel and u-boot and all the other shit with Debian stretch. Because I don't understand why NTC is lacking one. Anyway.